Working Compass tasks.
I don't know anything about the workflow and really, don't know anything
about Ruby or how these Gem in a jar things work, so someone who
actually uses this stuff is going to have to clean this up so that it
makes sense. I can't polish this for production myself but at least it's
functioning.